requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
altair==4.2.2
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
backports-zoneinfo==0.2.1
bcrypt==4.0.1
blinker==1.6.2
ca-certificates==2023.01.10
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
clearml==1.10.3
click==8.1.3
cycler==0.11.0
decorator==5.1.1
deta==1.1.0
docstring-parser==0.15
efficientnet-pytorch==0.7.1
entrypoints==0.4
extra-streamlit-components==0.1.56
flatbuffers==23.3.3
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
furl==2.1.3
gast==0.4.0
gitdb==4.0.10
gitpython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.11.0
kiwisolver==1.4.4
ld_impl_linux-64==2.38
libclang==16.0.0
libffi==3.4.2
libgcc-ng==11.2.0
libgomp==11.2.0
libstdcxx-ng==11.2.0
lightning-utilities==0.8.0
markdown==3.4.3
markdown-it-py==2.2.0
markupsafe==2.1.2
matplotlib==3.5.3
mdurl==0.1.2
multidict==6.0.4
munch==2.5.0
ncurses==6.4
networkx==2.6.3
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
openssl==1.1.1t
opt-einsum==3.3.0
